pyproject.toml
[project]
name = "convert-to-markdown"
version = "0.1.0"
description = "A simple wrapper for the MarkItDown library that provides command-line and macOS Quick Action integration for converting various file types to Markdown"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markitdown>=0.0.1a4",
    "openai>=1.64.0",
    "python-dotenv>=1.0.1",
    "uv>=0.6.3",
]

[project.scripts]
convert-to-md = "convert_to_md:main"
